OPENEUROPA-1922: Tests for the local translations importer.
