+++ docker-compose.yml
@@ -21,6 +21,23 @@ services:
 #    ports:
 #      - 3306:3306
 
+  selenium:
+    image: selenium/standalone-chrome-debug:3
+    expose:
+    - '4444'
+    environment:
+    - DISPLAY=:99
+    - SE_OPTS=-debug
+    - SCREEN_WIDTH=1280
+    - SCREEN_HEIGHT=800
+    - VNC_NO_PASSWORD=1
+    ports:
+    - '4444:4444'
+    - "5900:5900"
+    volumes:
+    - /dev/shm:/dev/shm
+    shm_size: 2g
+
 #### Mac users: uncomment the "volumes" key to enable the NFS file sharing. You can find more information about Docker for Mac here: https://github.com/openeuropa/openeuropa/blob/master/docs/starting/tooling.md#using-docker-on-macos
 
 #volumes:
