+++ tests/modules/translation_importer_test/translations/translation_importer_test-fr.po
@@ -1,0 +1,8 @@
+
+msgid ""
+msgstr ""
+
+msgid "Structure"
+msgstr "French STR"
+
+
